refactor: Sections 클래스 분리 + 중복된 의미의 메서드 통합
* 클래스 분리
1. SectionsValidator : Sections의 벨리데이션을 담당.
