+++ src/test/java/nextstep/subway/domain/SectionTest.java
@@ -2,6 +2,7 @@ package nextstep.subway.domain;
 
 import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException;
 
+import nextstep.subway.domain.Section.Section;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
